pyproject.toml
[project]
name = "django-app"
version = "0.1.0"
requires-python = ">3.7"
readme = "README.md"
authors = [{ name = "Team", email = "team@team.com" }]

dependencies = [
  # Django libraries
  "django",
  "djangorestframework",
  "gunicorn",
  # Database libraries
  "pgvector",
  "psycopg[binary]",
  "sqlalchemy",
  # Formatting
  "black",
  # Type checking
  "mypy",
  # Linting
  "ruff",
  # phidata
  "phidata==2.0.18",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]

# Change this value to use a different directory for the phidata workspace.
# [tool.phidata]
# workspace = "workspace"

[tool.ruff]
line-length = 110
exclude = ["appenv*", ".venv*"]
[tool.ruff.per-file-ignores]
# Ignore `F401` (import violations) in all `__init__.py` files
"__init__.py" = ["F401"]

[tool.mypy]
check_untyped_defs = true
no_implicit_optional = true
warn_unused_configs = true
plugins = ["pydantic.mypy"]
exclude = ["appenv*", ".venv*"]

[[tool.mypy.overrides]]
module = ["django.*", "pgvector.*", "setuptools.*"]
ignore_missing_imports = true
